Move Tailwind stateLayer plugin to its own file
