+++ src/config/tailwind/index.ts
@@ -1,2 +1,3 @@
 export * from './animations';
 export * from './autofillOverride';
+export * from './stateLayer';
