+++ tailwind.config.ts
@@ -1,40 +1,15 @@
 import type { Config } from 'tailwindcss';
 import scrollbar from 'tailwind-scrollbar';
 import animate from 'tailwindcss-animate';
-import plugin from 'tailwindcss/plugin';
-//@ts-expect-error Javascript only, no type decleration
-import flattenColorPalette from 'tailwindcss/lib/util/flattenColorPalette';
 
 // importing '@/config/tailwind' doesn't work here
-import { animations, autofillOverride } from './src/config/tailwind';
+import {
+  animations,
+  autofillOverride,
+  stateLayer,
+} from './src/config/tailwind';
 
 const shades = [50, ...[...Array(9).keys()].map(key => (key + 1) * 100), 950];
-
-const stateLayer = plugin(({ addUtilities, matchUtilities, theme }) => {
-  const themeColors = flattenColorPalette(theme('colors'));
-  addUtilities({
-    '.state-layer': {
-      position: 'relative',
-      overflow: 'hidden',
-      '&.fixed': { position: 'fixed' },
-      '&.absolute': { position: 'absolute' },
-      '&::after': {
-        content: '""',
-        position: 'absolute',
-        insetBlockStart: '0',
-        insetInlineStart: '0',
-        width: '100%',
-        height: '100%',
-        zIndex: '1',
-        transition: 'background-color 75ms cubic-bezier(0.4, 0, 0.2, 1)',
-      },
-    },
-  });
-  matchUtilities(
-    { 'state-layer': value => ({ '&::after': { backgroundColor: value } }) },
-    { values: themeColors, type: 'color' }
-  );
-});
 
 const config = {
   content: ['./src/**/*.{ts,tsx}'],
